FixTypeError

+++ FUNewsManagementSystem.BLL/Interfaces/ISystemAccountService.cs
@@ -10,7 +10,7 @@ namespace FUNewsManagementSystem.BLL.Interfaces
     public interface ISystemAccountService
     {
         Task<IEnumerable<SystemAccount>> GetAllAccountsAsync();
-        Task<SystemAccount?> GetAccountByIdAsync(int id);
+        Task<SystemAccount?> GetAccountByIdAsync(short id);
         Task AddAccountAsync(SystemAccount account);
         Task UpdateAccountAsync(SystemAccount account);
         Task DeleteAccountAsync(int id);
